Overriding python3 default behavior of printing floats, so that integers are printed.
One of the steps of creating input files for aaml is to indicate the sequence length. However, that sequence length cannot contain a decimal. This poses a problem because the default behavior of python3 is to print numbers as floats. This change overrides the default behavior so that an integer is printed instead.